requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
google-auth==2.25.2
python-dotenv==1.0.0
pytz==2024.1


